Restructured SDK to support device specific classes with high level functions of the device handled. Added Advanced Navigation Device for inheriting generic product features. Added Spatial class for Spatial device.
