+++ README.md
@@ -1,17 +1,6 @@
-Advanced Navigation Packet Protocol Spatial SDK
-
-Python Language SDK Version 0.02
+Advanced Navigation Python SDK
 
 Setup:
-- This SDK requires Python3.8 and above.
-- Add spatial_sdk to the Python path.
-- Example files have been made using:
-	- PySerial 3.4
-		- Users will need to install 3.4 and add to Python path.
-
-Change Log:
-- 0v02:
-	- Removed external libraries from SDK.
-	- Added Copyright header to files.
-- 0v01:
-	- First release.
+- This SDK requires Python 3.8 or greater and PySerial 3.4 or greater.
+- Add this directory to the Python path.
+- Add PySerial to the Python path.
