+++ examples/encode_decode_packet/packet_example.py
@@ -1,7 +1,7 @@
 ################################################################################
 ##                                                                            ##
-##                 Advanced Navigation Packet Protocol Library                ##
-##                  Python Language Spatial SDK, Version 0.02                 ##
+##                   Advanced Navigation Python Language SDK                  ##
+##                              packet_example.py                             ##
 ##                     Copyright 2021, Advanced Navigation                    ##
 ##                                                                            ##
 ################################################################################
@@ -27,32 +27,17 @@
 # DEALINGS IN THE SOFTWARE.                                                    #
 ################################################################################
 
+""" This example shows how to send, receive and log ANPP packets from a Spatial"""
 
-""" This example shows how to send, receive and log Spatial ANPP packets """
-
-import os
 import datetime
 import sys
 import math
 
-from spatial_sdk.spatial_packets import SensorRangesPacket, AccelerometerRange, GyroscopeRange, MagnetometerRange, SystemStatePacket, RawSensorsPacket, PacketID
-from spatial_sdk.an_packet_protocol import AN_Decoder, AN_Packet, an_packet_decode
-import serial
-import serial.serialutil as serialutil
+from spatial_device import Spatial
+from anpp_packets.an_packet_protocol import AN_Packet, an_packet_decode
 
 TRUE = 1
 FALSE = 0
-
-# Example function encoding an ANPP packet to be sent.
-def set_sensor_ranges():
-    sensor_ranges_packet = SensorRangesPacket()
-
-    sensor_ranges_packet.permanent = TRUE
-    sensor_ranges_packet.accelerometers_range = AccelerometerRange.accelerometer_range_4g.value
-    sensor_ranges_packet.gyroscopes_range = GyroscopeRange.gyroscope_range_500dps.value
-    sensor_ranges_packet.magnetometers_range = MagnetometerRange.magnetometer_range_8g.value
-
-    return(sensor_ranges_packet.encode().bytes())
 
 if __name__ == '__main__':
     # Checks enough arguments in command for serial communications. Otherwise prompts user on use.
@@ -62,60 +47,43 @@ if __name__ == '__main__':
     comport = str(sys.argv[1])
     baudrate = sys.argv[2]
 
-    # Checks if operating system is Windows or Linux
-    if ((os.name == 'nt') or (os.name == 'posix')):
-        # Connects to serial port
-        ser = serial.Serial(port=comport,
-                         baudrate=baudrate,
-                         bytesize=serialutil.EIGHTBITS,
-                         parity=serialutil.PARITY_NONE,
-                         stopbits=serialutil.STOPBITS_ONE,
-                         timeout=None,
-                         xonxoff=False,
-                         rtscts=False,
-                         write_timeout=None,
-                         dsrdtr=False,
-                         inter_byte_timeout=None
-                         )
-    else:
-        print(f"Packet_example currently only supports Windows and Linux operating systems.")
-        exit()
+    spatial = Spatial(comport, baudrate, log = True)
+    spatial.start_serial()
 
     # Checks serial port connection is open
-    if ser.is_open:
-        print(f"Connected to port:{comport} with baud:{baudrate}")
-        ser.flush()
+    if spatial.is_open:
+        print(f"Connected to port:{spatial.port} with baud:{spatial.baud}")
+        spatial.flush()
 
         # Creates log file for received binary data from device
         now = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
-        filename = ('SpatialLog_%s.anpp' %now)
-        logFile = open(filename, 'xb')
+        logFile = open(f"SpatialLog_{now}.anpp", 'xb')
 
-        decoder_array = AN_Decoder()
-        byteIndex = 0
         an_packet = AN_Packet()
 
-        system_state_packet = SystemStatePacket()
-        raw_sensor_packet = RawSensorsPacket()
+        # Sets sensor ranges
+        spatial.set_sensor_ranges(TRUE,
+                                  spatial.AccelerometerRange.accelerometer_range_4g.value,
+                                  spatial.GyroscopeRange.gyroscope_range_500dps.value,
+                                  spatial.MagnetometerRange.magnetometer_range_8g.value)
 
-        # Sets sensor ranges as an example of sending an ANPP packet
-        ser.write(set_sensor_ranges())
+        spatial.get_device_and_configuration_information()
 
-        while(1):
-            if (ser.in_waiting > 0):
+        while(spatial.is_open):
+            if (spatial.in_waiting() > 0):
                 # Get bytes in serial buffer
-                bytes_in_buffer = ser.in_waiting
-                data_bytes = ser.read(bytes_in_buffer)
+                bytes_in_buffer = spatial.in_waiting()
+                data_bytes = spatial.read(bytes_in_buffer)
 
                 # Record in log file the raw binary of ANPP packets
                 logFile.write(data_bytes)
 
                 # Adds bytes to array for decoding
-                decoder_array.add_data(packet_bytes = data_bytes)
+                spatial.bytes_waiting.add_data(packet_bytes = data_bytes)
 
             # If bytes are in array then decode
-            if (len(decoder_array.buffer) > 0):
-                an_packet, decoder_array = an_packet_decode(decoder_array)
+            if (len(spatial.bytes_waiting.buffer) > 0):
+                an_packet, spatial.bytes_waiting = an_packet_decode(spatial.bytes_waiting)
  
                 #===============================================================
                 # This example is only looking for two ANPP packets and printing a subset 
@@ -124,7 +92,8 @@ if __name__ == '__main__':
                 # Packet ID and Length printed in this example.
                 #===============================================================
                 if (an_packet is not None):
-                    if(an_packet.id == PacketID.system_state.value):
+                    if(an_packet.id == spatial.PacketID.system_state.value):
+                        system_state_packet = spatial.SystemStatePacket()
                         if (system_state_packet.decode(an_packet) == 0):
                             print(f"System State Packet:\n"
                                 f"\tLatitude:{math.degrees(system_state_packet.latitude)}, "
@@ -133,19 +102,34 @@ if __name__ == '__main__':
                             print(f"\tRoll:{math.degrees(system_state_packet.orientation[0])}, "
                                 f"Pitch:{math.degrees(system_state_packet.orientation[1])}, "
                                 f"Heading:{math.degrees(system_state_packet.orientation[2])}")
-                    elif(an_packet.id == PacketID.raw_sensors.value):
+                    elif(an_packet.id == spatial.PacketID.raw_sensors.value):
+                        raw_sensor_packet = spatial.RawSensorsPacket()
                         if(raw_sensor_packet.decode(an_packet) == 0):
-                            print(f"Raw Sensors Packet: \n"
+                            print(f"Raw Sensors Packet:\n"
                                 f"\tAccelerometers X:{raw_sensor_packet.accelerometers[0]}, "
                                 f"Y:{raw_sensor_packet.accelerometers[1]}, "
                                 f"Z:{raw_sensor_packet.accelerometers[2]}")
                             print(f"\tGyroscopes X:{math.degrees(raw_sensor_packet.gyroscopes[0])}, "
                                 f"Y:{math.degrees(raw_sensor_packet.gyroscopes[1])}, "
                                 f"Z:{math.degrees(raw_sensor_packet.gyroscopes[2])}")
+                    elif(an_packet.id == spatial.PacketID.external_air_data.value):
+                        external_air_data_packet = spatial.ExternalAirDataPacket()
+                        if(external_air_data_packet.decode(an_packet) == 0):
+                            print(f"External Air Data Packet:\n"
+                                  f"\tBarometric Altitude Set and Ready:{external_air_data_packet.flags.barometric_altitude_set_and_valid}\n "
+                                  f"\tBarometric Altitude: {external_air_data_packet.barometric_altitude}m\n"
+                                  f"\tAirspeed Set and Ready:{external_air_data_packet.flags.airspeed_set_and_valid}\n"
+                                  f"\tAirspeed: {external_air_data_packet.airspeed}m/s")
+                    elif(an_packet.id == spatial.PacketID.external_odometer.value):
+                        external_odometer_packet = spatial.ExternalOdometerPacket()
+                        if(external_odometer_packet.decode(an_packet) == 0):
+                            print(f"External Odometer Packet:\n"
+                                  f"\tSpeed:{external_odometer_packet.speed}m/s\n"
+                                  f"\tDistance Travelled:{external_odometer_packet.distance_travelled}m\n"
+                                  f"\tReverse Detection Supported:{external_odometer_packet.flags.reverse_detection_supported}")
                     else:
                         print(f"Packet ID:{an_packet.id} of Length:{an_packet.length}")
-
     else:
-        print(f"No connection. :( ")
+        print(f"No connection.")
 
-    ser.close()
+    spatial.close()
